PLEASE GRADE THE LATEST ONE PLEASE!!!
I know it's way overdue but i haven't even added functionality and it won't take me more than an hour to do the probably easiest part which is adding functionality. Please I beg you.

Got the string finding function to mostly work, now I just need to combine it into the token finder to make it simpler and also to be able to deal with spaces correctly.