done for now PLEASE GRADE OR AT LEAST CONSIDER UP TO HERE I BEG PLEASE
I got a little bit of the scaffolding in for implementing functionality, I just need to think through how to combine the stack and the code, along with figuring out how to assign to functions (e.g, setting a string to a read function). It is capible of printing, and can be tested on a hello world-esque script.